feat. 관리자 컴포넌트 정리
